backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.22
asyncpg>=0.28.0
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Redis
redis>=5.0.1
aioredis>=2.0.1

# Audio Processing
librosa>=0.10.1
numpy>=1.24.0
pandas>=2.1.2
scikit-learn>=1.3.2

# Feature Storage
faiss-cpu>=1.7.4

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.2.3
httpx==0.28.1
psutil>=6.0.0

# Additional packages
nltk>=3.6.0
gensim>=4.0.0
textblob>=0.15.3
vaderSentiment>=3.3.2
kagglehub>=0.1.0
yt-dlp>=2023.12.30
lyricsgenius>=3.0.1
requests>=2.31.0
beautifulsoup4>=4.12.0
youtube-search-python==1.6.2 